refactor: separate API calls into files
